refactor(criteria): split Criteria into AbstractCriteria + concrete Criteria
Extracts the pure tree-structure fields/behavior (name, parent/root/
ascendants, tree constraints, save/validation) into an abstract
AbstractCriteria(PrivateUniqueResource) class that has no knowledge of
UploadedTrack. Criteria becomes Criteria(AbstractCriteria,
UploadedTrackMixin), keeping only the uploaded-track-coupled bits
(uploaded_tracks property, CriteriaManager, concrete Meta/db_table).

Proves the diamond-inheritance shape works entirely within hear-api
before any shared package is introduced. No behavior change.
