+++ api/model/criteria/Criteria.py
@@ -1,78 +1,27 @@
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
-from django.db import IntegrityError, models
-from django.db.models import QuerySet
-from django.utils.translation import gettext as _
+from django.db import models
 
-from api import settings
-from api.exception.validation.app.AppValidationException import AppValidationException
-from api.exception.validation.FieldValidationErrorCode import FieldValidationErrorCode
+from api.model.criteria.AbstractCriteria import AbstractCriteria
 from api.model.criteria.CriteriaManager import CriteriaManager
-from api.model.criteria.lineage_rel.Fields import Fields as CriteriaLineageRelFields
-from api.model.field.AppCharField import AppCharField
-from api.model.field.foreign_key.AppForeignKey import AppForeignKey
-from api.model.field.foreign_key.PrivateForeignKey import PrivateForeignKey
-from api.model.field.foreign_key.PrivateManyToManyField import PrivateManyToManyField
 from api.model.uploaded_track_mixin.UploadedTrackMixin import UploadedTrackMixin
-from api.utils.model import SaveContext
 
 from .Fields import Fields
-from .type.CriteriaType import CriteriaType
 
 if TYPE_CHECKING:
     from api.model.playlist.children.criteria.CriteriaPlaylist import CriteriaPlaylist
     from api.model.uploaded_track.UploadedTrack import UploadedTrack
 
-    from .lineage_rel.CriteriaLineageRel import CriteriaLineageRel
 
-
-class Criteria(UploadedTrackMixin):
-    _name = AppCharField(max_length=settings.CRITERIA_NAME_LEN_MAX, db_column=Fields.NAME_PUBLIC)
-    ascendants: QuerySet[Criteria] = PrivateManyToManyField(
-        "self",
-        through="CriteriaLineageRel",
-        through_fields=(CriteriaLineageRelFields.DESCENDANT, CriteriaLineageRelFields.ASCENDANT),
-        symmetrical=False,
-    )  # type: ignore
-    parent: Criteria | None = PrivateForeignKey(
-        "self", on_delete=models.SET_NULL, null=True, related_name=Fields.CHILDREN
-    )  # type: ignore
-
-    root: Criteria = PrivateForeignKey("self", on_delete=models.DO_NOTHING, related_name=Fields.DESCENDANTS)  # type: ignore
-
-    type = AppForeignKey(CriteriaType, on_delete=models.CASCADE)
-
+class Criteria(AbstractCriteria, UploadedTrackMixin):
     if TYPE_CHECKING:
-        ascendants_rels: QuerySet[CriteriaLineageRel]
-        descendants: QuerySet[Criteria]
-        descendants_rels: QuerySet[CriteriaLineageRel]
-        children: QuerySet[Criteria]
         criteria_playlist: CriteriaPlaylist
 
     objects: CriteriaManager = CriteriaManager()
 
     @property
-    def name(self) -> str:
-        return self._name
-
-    @property
-    def uploaded_tracks(self) -> models.QuerySet[UploadedTrack]:
+    def uploaded_tracks(self) -> models.QuerySet["UploadedTrack"]:
         return getattr(self, Fields.UPLOADED_TRACKS_RELATED_NAME)
-
-    @property
-    def is_root(self) -> bool:
-        return not self.parent
-
-    @property
-    def descendant_list(self) -> list[Criteria]:
-        """
-        Get all descendants of this criteria using the lineage system.
-        This is more efficient than recursive traversal as it uses the pre-computed relationships.
-
-        Returns:
-            A list of all descendant criteria
-        """
-        return list(self.descendants.all())
 
     class Meta:
         db_table = "htmt_api_criteria"
@@ -86,59 +35,3 @@ class Criteria(UploadedTrackMixin):
             models.Index(fields=[Fields.USER, Fields.NAME_INTERNAL], name="%(class)s_user_name_idx"),
             models.Index(fields=[Fields.USER, Fields.UUID], name="%(class)s_user_uuid_idx"),
         ]
-
-    def __str__(self) -> str:
-        parent_str = f"{Fields.PARENT}: {self.parent.name}" if self.parent else f"[no {Fields.PARENT}]"
-        created_on_str = f"{Fields.CREATED_ON}: {self.created_on}"
-        updated_on_str = f"{Fields.UPDATED_ON}: {self.updated_on}"
-
-        return f"{self.uuid} | {self.name} | {parent_str} | {created_on_str} | {updated_on_str}"
-
-    def _set_root(self):
-        current_root = getattr(self, f"{Fields.ROOT}", None)
-        new_root = self.parent.root if self.parent else self
-
-        new_root_pk = None
-        if not new_root:
-            new_root_pk = self.pk
-        elif current_root != new_root:
-            new_root_pk = new_root.pk
-
-        if new_root_pk:
-            self.root_id = new_root_pk
-            return True
-        return False
-
-    def _prepare_save(self, ctx: SaveContext) -> dict:
-        self._set_uuid_if_necessary()
-        root_has_changed = self._set_root()
-        if not self._state.adding and root_has_changed:
-            ctx.add_modified_field(f"{Fields.ROOT}_id")
-        return ctx.kwargs
-
-    def save(self, *args: Any, **kwargs: Any) -> None:
-        try:
-            super().save(*args, **kwargs)
-        except IntegrityError as e:
-            error_message = str(e)
-            if "non_empty_name" in error_message:
-                raise AppValidationException(
-                    field_name=Fields.NAME_PUBLIC,
-                    message=_("Name cannot be empty"),
-                    field_validation_error_code=FieldValidationErrorCode.NAME_EMPTY,
-                )
-            if "unique_name_per_user" in error_message:
-                raise AppValidationException(
-                    field_name=Fields.NAME_PUBLIC,
-                    message=_(f'The name "{self.name}" is already used'),
-                    field_validation_error_code=FieldValidationErrorCode.NAME_DUPLICATE,
-                )
-            # Let other database integrity errors propagate to be handled as system errors
-            raise
-
-    def is_descendant_of(self, other_criteria: Criteria) -> bool:
-        if self.parent == other_criteria:
-            return True
-        if self.parent:
-            return self.parent.is_descendant_of(other_criteria)
-        return False
